JMP-34: Rename template-service to order-service and update related configurations.
