+++ .env.example
@@ -2,7 +2,7 @@
 DEBUG=true
 
 # Database
-DB_NAME=template-service_db
+DB_NAME=order-service_db
 DB_HOST=localhost
 DB_PORT=5432
 DB_USER=postgres
