+++ settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "template-service"
+rootProject.name = "order-service"
