+++ sonar-project.properties
@@ -1,4 +1,4 @@
-sonar.projectKey=Java-Marketplace_template
+sonar.projectKey=Java-Marketplace_order-service
 sonar.organization=java-marketplace
 
 sonar.sources=./src/main
