+++ src/main/resources/application.yaml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: template-service
+    name: order-service
 
   datasource:
     url: jdbc:postgresql://${DB_HOST}:${DB_PORT}/${DB_NAME}
